Adding an -updater:willBeginUpdate: SUUpdaterDelegate method.
This method is called just before the update process starts
i.e. before downloading or installing.

The delegate may return NO to cancel the update process.
This is useful in some cases e.g. if the app needs to check whether
other apps are running before continuing.

Doing so earlier in the update process is un-necessarily invasive and
may cause the user to close running apps (save work etc) when an
update is not even available.